requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
pydantic
langchain
langchain-google-genai
httpx
python-dotenv
langchain-core
pydantic-settings
python-telegram-bot==20.7
